Refactor Styles.js to add red color and buttonText style Update Activities.js to navigate to AddEditActivity instead of EditActivity Refactor EditActivity.js to add PressableButton component and cancel button functionality Update App.js to navigate to AddEditActivity instead of AddActivity Delete AddActivity.js
