+++ App.js
@@ -10,9 +10,8 @@ import SimpleLineIcons from '@expo/vector-icons/SimpleLineIcons';
 import Activities from './screens/Activities';
 import Diet from './screens/Diet';
 import Settings from './screens/Settings';
-import AddActivity from './screens/AddActivity';
+import AddEditActivity from './screens/AddEditActivity';
 import AddDiet from './screens/AddDiet';
-import EditActivity from './screens/EditActivity';
 import { ItemsProvider } from './components/ItemsContext';
 import { ThemeProvider } from './components/ThemeContext';
 import Styles, { commonHeaderStyles, commonBottomTabStyles, commonStyles } from './components/Styles';
@@ -49,7 +48,7 @@ export default function App() {
           options={({ navigation }) => ({
             headerRight: () => (
               <PressableButton
-                pressedFunction={() => navigation.navigate('AddActivity')}
+                pressedFunction={() => navigation.navigate('AddEditActivity')}
               >
               <View style={commonStyles.headerButtonContainer}>
                 <AntDesign name="plus" size={24} color="white" />
@@ -59,26 +58,26 @@ export default function App() {
             ),
           })}
         />
-        <Stack.Screen 
-          name="AddActivity"
-          component={AddActivity}
-          options={{ title: 'Add An Activity' }} />
-        
+        {/* Add / Edit Activity Screen */}
         <Stack.Screen
-          name="EditActivity"
-          component={EditActivity}
-          options={({route, navigation}) => ({ 
-            title: 'Edit', 
-            headerRight: () => (
-              <PressableButton
-                pressedFunction={() => handleDelete(route.params.item.id, "activities", navigation)}
-              >
-              <View style={commonStyles.headerButtonContainer}>
-                <AntDesign name="delete" size={24} color="white" />
-              </View>
-              </PressableButton>
-            ),
-            })} />
+          name="AddEditActivity"
+          component={AddEditActivity}
+          options={({ route, navigation }) => ({
+            title: route.params?.activity ? 'Edit Activity' : 'Add Activity', // Conditional title
+            headerRight: () =>
+              route.params?.activity && ( // Only show delete button for editing
+                <PressableButton
+                  pressedFunction={() =>
+                    handleDelete(route.params.activity.id, 'activities', navigation)
+                  }
+                >
+                  <View style={commonStyles.headerButtonContainer}>
+                    <AntDesign name="delete" size={24} color="white" />
+                  </View>
+                </PressableButton>
+              ),
+          })}
+        />
       </Stack.Navigator>
     );
   }
